+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 - MVP-404 - make all tests use PostgreSQL
 - MVP-380 - add UI test to add multiple countries on target markets page
 - MVP-405 - create initial export plan
+- MVP-407 - refactor Browser tests
 - no-ticket - update export-plan link
 - MVP-406 - hook up remove country